move con decoder to layers
